+++ NEWS.md
@@ -1,16 +1,28 @@
-# xplainfi 0.1.0.9000
+# xplainfi 0.1.0.9001
 
 This turns out to be still a period of major changes in the early phase, so, uhm, well.
 
 ## General changes and improvements
 
-- `$importance` become a function `$importance()` with arguments `standardize` and `variance_method` (#40):
-  - `"nadeau_bengio"` implements the correction method by Nadeau & Bengio (2003) recommended by Molnaet et al. (2023).
-- Add `$obs_loss` and `$predictions` fields to `FeatureImportanceMeasure`, now used by `LOCO` and `LOCI`
-  - Both get arugments `obs_loss = FALSE` use the measure's `$aggregator` for aggregation in case of `obs_loss = TRUE`, to allow for  median of absolute differences calculation as in original LOCO formulation, rather than the "micro-"averaged approach calculated by default.
+- `$importance` becomes a function `$importance()` with arguments `standardize` and `variance_method` (#40):
+  - `"nadeau_bengio"` implements the correction method by Nadeau & Bengio (2003) recommended by Molnar et al. (2023).
+- `$scores` becomes `$scores()` for more fleixibility analogous to `$importance()`. Original scores are stored in `private$.scores` (for now).
+- Both `$importance()` and `$scores()` have argument `relation` defaulting to `"differrence"` which allows to calculate importances like PFI as either the difference or the ratio of baseline and post-modification loss. The argument is moved out of `$compute()` to avoid having to recompute any predictions or model refits.
 - Add `sim_dgp_ewald()` and other `sim_dgp_*()` helpers to simulate data (in `Task` form) with simple DGPs as used for illustration in Ewald et al. (2024) for example, which should make it easier to interpret the results of various importance methods.
 
+### Observation-wise losses
+
+-  `$obs_loss()` analogously to `$scores()` computes observation-wise importance scores based on losses stored in `$.obs_losses` **if** `measure` has a `Measure$obs_loss()` function.
+
+- `$predictions` will probably be removed again?
+
 ## Method-specific changes
+
+### `LeaveOutIn` -> `WVIM`
+
+- Williamson's Variable Importance Measure (WVIM) generalizes LOCO / LOCI
+- Feature grouping API is still WIP
+- New implementation built around `mlr3fselect`, greatly simplifying the internals
 
 ### `PerturbationImportance`
 
@@ -27,7 +39,7 @@ This turns out to be still a period of major changes in the early phase, so, uhm
 ### `SAGE`
 
 - Fix accidentally marginal `ConditionalSAGE`.
--  Also using `learner$predict_newdata_fast()` now
+- Also using `learner$predict_newdata_fast()` now  (#39)
 - `batch_size` controls number of observations used at once per `learner$predict_newdata_fast()` call (could lead to excessive RAM usage). 
 - convergence tracking if `early_stopping = TRUE` ([#29](https://github.com/jemus42/xplainfi/pull/29))
   - Permutations are evaluated in steps of `check_interval` at a time, after each convergence is checked